requirements.txt
flask>=3.0.0,<4.0.0
reportlab>=4.0.0,<5.0.0
pdfplumber>=0.11.0,<0.12.0
python-docx>=1.1.0,<2.0.0
pytest>=8.0.0,<9.0.0
requests>=2.31.0,<3.0.0
